+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   * Role now uses active\_attr
   * RegisterField now uses active\_attr
   * RegisterData now uses active\_attr
+  * Member now uses active\_attr
 
 ## Version 0.0.17